pdf 업로드 수정

+++ backend/src/main/java/com/gigigenie/domain/chat/repository/EmbeddingRepository.java
@@ -3,4 +3,5 @@ package com.gigigenie.domain.chat.repository;
 import com.gigigenie.domain.chat.entity.Embedding;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmbeddingRepository extends JpaRepository<Embedding, Long> {}
+public interface EmbeddingRepository extends JpaRepository<Embedding, Integer> {
+}
